Integrate pathway scoring into batch runner
